Refactored key, simple responses working
