+++ src/main/java/com/postman/collection/PostmanAPIKey.java
@@ -1,11 +1,10 @@
 package com.postman.collection;
 
 import com.google.gson.*;
-
+import java.util.UUID;
 public class PostmanAPIKey implements IPostmanCollectionElement {
     private String type = "";
     private String value = "";
-    private String key = "";
     private transient String key = UUID.randomUUID().toString();
 
     public PostmanAPIKey(String type, String value, String key) {
