+++ src/main/java/com/postman/collection/PostmanEvent.java
@@ -86,11 +86,7 @@ public String getKey() {
     return listen;
 }
 
-@Override
-public void setKey(String key) {
-    
-    
-}
+
 
 @Override
 public String toJson(boolean escaped, enumVariableResolution variableStrategy) {
